fix: move /new homepage to /
