--- packages/paste-website/src/components/homepage/GetStartedCardIllustration.tsx
@@ -1,6 +1,0 @@
-import { Box } from "@twilio-paste/box";
-import * as React from "react";
-
-export const GetStartedCardIllustration: React.FC<React.PropsWithChildren> = ({ children }) => {
-  return <Box marginBottom="space70">{children}</Box>;
-};
