+++ packages/paste-website/src/components/site-wrapper/site-footer/SiteFooterHeader.tsx
@@ -1,8 +1,7 @@
 import { Box } from "@twilio-paste/box";
-import { Text } from "@twilio-paste/text";
 import { useTheme } from "@twilio-paste/theme";
 
-import HeartDoodleIcon from "../../icons/HeartDoodleIcon";
+import HeartDoodleIcon from "../../icons/NewHeartDoodleIcon";
 
 const SiteFooterHeader = (): JSX.Element => {
   const theme = useTheme();
@@ -19,18 +18,6 @@ const SiteFooterHeader = (): JSX.Element => {
         color={theme.textColors.colorTextInverse}
         decorative
       />
-      <Text
-        as="h2"
-        color="colorTextInverse"
-        fontFamily="fontFamilyDisplay"
-        fontSize={["fontSize60", "fontSizeDisplay10"]}
-        fontWeight="fontWeightExtrabold"
-        lineHeight={["lineHeight60", "lineHeightDisplay10"]}
-        marginLeft={["space0", "space40"]}
-        marginTop={["space40", "space0"]}
-      >
-        Black lives matter.
-      </Text>
     </Box>
   );
 };
