+++ packages/paste-website/src/components/site-wrapper/site-header/index.tsx
@@ -46,8 +46,8 @@ export const SiteHeader: React.FC<{
             <SiteHeaderSearch />
           </Box>
           <AssistantLink />
-          <DarkModeToggle />
           <ContactUsMenu />
+          <DarkModeToggle />
           <FigmaButton />
           <GitHubButton type="stargazers" namespace="twilio-labs" repo="paste" />
         </TopbarActions>
